15_Open_Source_Endpoints/Application_Start/requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-core
langchain-huggingface
faiss-cpu
tqdm
python-dotenv
sentence-transformers